refactor: move github/cli to gh package
