+++ gh/doc.go
@@ -1,0 +1,3 @@
+// Package gh provides a library for CLI Go applications to help interface with the go-gh CLI tool which is
+// a go library for interfacing with the GitHub CLI tool.
+package gh
