--- github/cli/doc.go
@@ -1,3 +1,0 @@
-// Package cli provides a library for CLI Go applications to help interface with the go-gh CLI tool which is
-// a go library for interfacing with the GitHub CLI tool.
-package cli
