--- runner/jobs/doc.go
@@ -1,2 +1,0 @@
-// Package jobs contains necessary types and functions to orchestrate jobs executions as part of a workflow run.
-package jobs
