--- runner/state/container.go
@@ -1,9 +1,0 @@
-package state
-
-import "dagger.io/dagger"
-
-type ContainerState struct {
-	BaseState
-
-	Container *dagger.Container
-}
