--- runner/steps/doc.go
@@ -1,2 +1,0 @@
-// Package steps contains necessary types and functions to orchestrate the steps of a job run.
-package steps
