--- runner/workflows/doc.go
@@ -1,2 +1,0 @@
-// Package workflows contains necessary types and functions to orchestrate workflows executions.
-package workflows
